Feature fix 2.0

+++ src/main/java/com/example/bulletinboard/service/impl/CommentServiceImpl.java
@@ -1,58 +1,84 @@
 package com.example.bulletinboard.service.impl;
 
 import com.example.bulletinboard.dto.*;
+import com.example.bulletinboard.entity.Ad;
 import com.example.bulletinboard.entity.Comment;
+import com.example.bulletinboard.entity.User;
+import com.example.bulletinboard.repository.AdRepo;
 import com.example.bulletinboard.repository.CommentRepo;
-import com.example.bulletinboard.service.AdService;
+import com.example.bulletinboard.repository.UserRepo;
 import com.example.bulletinboard.service.CommentMapper;
 import com.example.bulletinboard.service.CommentService;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 
-import java.time.Instant;
-import java.util.List;
-import java.util.Optional;
-import java.util.concurrent.atomic.AtomicReference;
-import java.util.stream.Collectors;
+import java.util.NoSuchElementException;
 
 @Service
 @RequiredArgsConstructor
+@Slf4j
 public class CommentServiceImpl implements CommentService {
     private final CommentRepo commentRepo;
     private final CommentMapper commentMapper;
-    private final AdService adService;
+    private final AdRepo adRepo;
+    private final UserRepo userRepo;
+    private final UserDetails userDetails;
 
     @Override
-    public CommentDto create(Integer id, CommentDto comment) {
-        return commentMapper.toDto(commentRepo.save(commentMapper.toComment(comment)));
+    public CommentDto create(Integer adId, CreateOrUpdateComment createComment) {
+        Ad ad = getAd(adId);
+        User user = getUser();
+        Comment comment = commentMapper.updateToComment(createComment);
+        comment.setUser(user);
+        comment.setAd(ad);
+        commentRepo.save(comment);
+        log.info("createComment");
+        return commentMapper.toDto(comment);
     }
 
     @Override
-    public Optional<CommentDto> getCommentsByAdId(Integer id) {
-        return Optional.ofNullable(commentMapper.toDto(commentRepo.findById(id).
-                orElse(null)));
-    }
-    @Override
-    public List<CommentDto> getAll() {
-        return commentRepo.findAll()
-                .stream()
-                .map(commentMapper::toDto)
-                .collect(Collectors.toList());
+    public Comments getAll(Integer adId) {
+        log.info("getAllComments");
+        return commentMapper.to(commentRepo.findAllByAd_Id(adId));
     }
 
     @Override
-    public void delete(Integer id) {
-        Comment comment = commentRepo.getById(id);
-        commentRepo.delete(comment);
+    public void delete(Integer adId, Integer commentId) {
+        User user = getUser();
+        Ad ad = getAd(adId);
+        Comment comment = commentRepo.deleteCommentByIdAndAd_Id(adId, commentId);
+        if (rightsVerification(user, ad)) {
+            commentRepo.delete(comment);
+        } else {
+            throw new UnsupportedOperationException("Нет прав на удаление комментария");
+        }
     }
     @Override
     public void updateComment(Integer commentId , CreateOrUpdateComment comment) {
-        AtomicReference<Optional<CreateOrUpdateComment>> atomicReference = new AtomicReference<>();
-        commentRepo.findById(commentId).ifPresentOrElse(foundComment -> {
+        commentRepo.findById(commentId).map(foundComment -> {
             foundComment.setText(comment.getText());
-            atomicReference.set(Optional.of(commentMapper.fromUpdateComment(commentRepo.save(foundComment))));
-        }, () -> {
-            atomicReference.set(Optional.empty());
-        });
+            return commentMapper.fromUpdateComment(commentRepo.save(foundComment));
+        }).orElseThrow(CommentNotFoundException::new);
+    }
+
+    private class CommentNotFoundException extends RuntimeException {
+        public CommentNotFoundException() {
+            super("Комментарий не найден");
+        }
+    }
+
+    private Ad getAd(Integer id) {
+        return adRepo.findById(id).orElseThrow(()-> new NoSuchElementException("Объявление не найдено")) ;
+    }
+
+    private User getUser() {
+        return userRepo.findByEmail(userDetails.getUsername()).orElseThrow(() -> new UsernameNotFoundException("Пользователь не найден"));
+    }
+
+    private boolean rightsVerification(User user, Ad ad) {
+        return (user.getRole().equals(Role.ADMIN) || ad.getUser().equals(user));
     }
 }
